Complete rewrite of alpha beta
Now the chosen depth is the ply-depth instead of the move-depth.
Longer searches but less logical problems.
Simu Players are now dummies instead of implementing and deriving methods.
Allows for a single method, which should be faster (has not been compared).
find-all-possible-moves has been replaced with find-valid-1moves, which searches
only for moves on the four cells around, and not farther.
When the best ply is found, it is saved to be played in one row,
instead of running the search again (which would be useless now).

But the computation time can still become extremely long, even on a small
5x5 board.